// src/models/OrderForm.ts
export interface OrderFormI {
  accessCode: string;
  city: string;
  companyName: string;
  deliveryTime: string;
  firstLastName: string;
  floor: string;
  email: string;
  flatNumber: string;
  houseNumber: string;
  payWith: string;
  phoneNumber: string;
  postcode: string;
  streetName: string;
  note: string;
  voutcher: string;
}